modernize - separate integration and unit tests
