+++ package.json
@@ -25,19 +25,21 @@
     "node": ">=8.1.2"
   },
   "devDependencies": {
-    "chai": "3.5.0",
+    "chai": "4.1.2",
     "chai-things": "0.2.0",
-    "coveralls": "2.11.11",
+    "coveralls": "3.0.0",
     "eslint": "4.16.0",
-    "istanbul": "0.4.4",
-    "mocha": "2.5.3 ",
-    "mockery": "1.7.0",
-    "sinon": "1.17.4"
+    "istanbul": "0.4.5",
+    "mocha": "5.0.0",
+    "mockery": "2.1.0",
+    "sinon": "4.2.2"
   },
   "scripts": {
-    "test": "mocha --recursive test/unit",
+    "test": "npm run test-unit && npm run test-integration",
+    "test-unit": "mocha --recursive test/unit",
+    "test-integration": "mocha --recursive test/integration",
     "lint": "./node_modules/.bin/eslint ./lib/ ./test ./index.js",
-    "fix-lint": "npm run lint -- --fix",
+    "lint-fix": "npm run lint -- --fix",
     "cover": "node ./node_modules/.bin/istanbul cover ./node_modules/mocha/bin/_mocha --include-all-sources --report lcovonly -- -R spec test/unit --recursive",
     "coveralls": "npm run cover && ./node_modules/coveralls/bin/coveralls.js < coverage/lcov.info"
   }
